feat(experiments): move sequential testing config into the stats modal
Sequential testing only applies to the frequentist engine, so its controls now live inside the Statistics configuration modal and appear when Frequentist is selected, instead of as a separate always-visible settings section with its own modal. The settings tab keeps a compact "· Sequential testing" indicator in the Statistics summary.
