+++ frontend/src/scenes/experiments/modalsLogic.ts
@@ -36,8 +36,6 @@ export const modalsLogic = kea<modalsLogicType>([
         closeStatsEngineModal: true,
         openCupedModal: true,
         closeCupedModal: true,
-        openSequentialTestingModal: true,
-        closeSequentialTestingModal: true,
         openPrimaryMetricModal: (uuid: string) => ({ uuid }),
         closePrimaryMetricModal: true,
         openSecondaryMetricModal: (uuid: string) => ({ uuid }),
@@ -198,13 +196,6 @@ export const modalsLogic = kea<modalsLogicType>([
             {
                 openCupedModal: () => true,
                 closeCupedModal: () => false,
-            },
-        ],
-        isSequentialTestingModalOpen: [
-            false,
-            {
-                openSequentialTestingModal: () => true,
-                closeSequentialTestingModal: () => false,
             },
         ],
         isRunningTimeConfigModalOpen: [
